coucou

+++ nucleo/src/asserv.hpp
@@ -3,7 +3,7 @@
 
 
 void conversion_d_alpha();
-uint8_t sature(uint8_t valeur, uint8_t min, uint8_t max);
+uint8_t sature(float valeur, uint8_t min, uint8_t max);
 void asserv_periodique();
 
 #endif
